fixed nanotron config file placed in a very random place
